backend/restapi/serializer.py
from rest_framework import serializers
from .models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ("name", "description", "department", "price", "image_url", "hovered_image_url")
